+++ pipelines/pipe1.R
@@ -3,8 +3,7 @@
 #' This is a function that runs every hour starting at
 #' 2024-03-01 09:00:00
 #'
-#' @maestroFrequency day
-#' @maestroInterval 1
+#' @maestroFrequency 1 day
 #' @maestroStartTime 2024-03-01 09:00:00
 #' @maestroTz UTC
 #'