+++ tests/testthat/test-build_schedule_entry.R
@@ -5,8 +5,9 @@ test_that("can create a schedule entry from a single well-documented fun", {
   expect_s3_class(res, "tbl_df")
   expect_equal(nrow(res), 1)
   expect_in(
-    c("script_path", "pipe_name", "frequency", "interval",
-      "start_time", "tz", "skip", "log_level"),
+    c("script_path", "pipe_name", "frequency",
+      "start_time", "tz", "skip", "log_level",
+      "frequency_n", "frequency_unit"),
     names(res)
   )
   expect_snapshot(res)