+++ tests/testthat/test_pipelines/test_pipeline_daily_default.R
@@ -1,7 +1,6 @@
 #' Simple mtcars print function
 #'
 #' @maestroFrequency
-#' @maestroInterval
 #' @maestroStartTime
 #' @maestroTz
 #'