+++ tests/testthat/test_pipelines_parse_all_bad/tagged_but_no_func.R
@@ -1,4 +1,4 @@
-#' @maestroFrequency day
+#' @maestroFrequency 1 day
 library(dplyr)
 
 data <- iris |>