+++ tests/testthat/test_pipelines_parse_all_bad/test_pipeline_daily_bad.R
@@ -1,7 +1,6 @@
 #' Simple mtcars print function
 #'
 #' @maestroFrequency 17
-#' @maestroInterval
 #' @maestroStartTime
 #'
 #' @export