+++ tests/testthat/test_pipelines_run_all_good/pipe1.R
@@ -3,10 +3,9 @@
 #' This is a function that runs every hour starting at
 #' 2024-03-01 09:00:00
 #'
-#' @maestroFrequency day
-#' @maestroInterval 1
+#' @maestroFrequency 1 day
 #' @maestroStartTime 2024-03-01 09:00:00
-#' @maestroTz UTC
+#' @maestroTz America/Halifax
 #'
 #' @export
 get_mtcars <- function() {
@@ -16,16 +15,14 @@ get_mtcars <- function() {
 
 #' Multiply
 #'
-#' @maestroFrequency month
-#' @maestroInterval 3
+#' @maestroFrequency 3 month
 #'
 #' @export
 wait <- function() {
   Sys.sleep(0.01)
 }
 
-#' @maestroFrequency week
-#' @maestroInterval 1
+#' @maestroFrequency 1 week
 #'
 #' @export
 weekly <- function() {
@@ -33,6 +30,7 @@ weekly <- function() {
 }
 
 #' @maestroStartTime 5000-12-12 10:10:10
+#' @maestroFrequency 30 seconds
 way_in_the_future <- function() {
   invisible()
 }